refactor(gui): ♻️ Extract launcher env GUI section from ConfigGuiBuilder
Move conda/pixi visit methods (~390 lines) into a new package-private
LauncherEnvGuiSection class. ConfigGuiBuilder delegates via two-line
stubs. Result: ConfigGuiBuilder 800 lines, LauncherEnvGuiSection 465 lines.
